fix: security parse id

+++ apps/user_profiles/entity/user_profiles.request.go
@@ -1,7 +1,7 @@
 package entity
 
 type UpdateUserRequestDTO struct {
-	ID        uint   `validate:"required,number,min=1"`
+	ID        uint64 `validate:"required,number,min=1"`
 	FirstName string `validate:"required,alpha,min=2"`
 	LastName  string `validate:"required,alpha,min=2"`
 	Phone     string `validate:"required,e164,min=11"`
