+++ apps/users/entity/users.response.go
@@ -6,7 +6,7 @@ import (
 )
 
 type UserResponse struct {
-	ID        int       `json:"id"`
+	ID        uint64    `json:"id"`
 	Email     string    `json:"email"`
 	Password  string    `json:"password"`
 	IsActive  *bool     `json:"is_active"`
@@ -14,7 +14,7 @@ type UserResponse struct {
 }
 
 type LoginUserResponse struct {
-	ID        uint                 `json:"id"`
+	ID        uint64               `json:"id"`
 	Email     string               `json:"email"`
 	Token     string               `json:"token"`
 	ExpiresAt string               `json:"expires_at"`
